Teach soong not to duplicate the HWASAN runtime into each APEX.
When HWASAN is enabled, the runtime is conceptually part of Bionic (and
mutually depends on it), so it needs to be treated in the same way as the
Bionic libs.

Now there are only two copies of the runtime: the one in
/system/lib64/bootstrap (which won't be used by ordinary processes) and the
one in the runtime APEX.

This reduces the size of the HWASAN system image and fixes an issue where
multiple copies of the HWASAN runtime were being loaded into 64-bit binaries in
APEXes because the linker namespace for the binary is different from the one
for its dependent libraries outside of APEXes. HWASAN only supports loading
one copy of the runtime per process, so this was causing such binaries to
crash on startup.

Bug: http://b/134503977
Test: Builds

Change-Id: I228896e193a035e6dfba9f6e28d0b2e12fc163ea
Merged-In: I228896e193a035e6dfba9f6e28d0b2e12fc163ea
(cherry picked from commit 3478bb2a7fbcaaea1326a1df4a564d7ca24056ce)