use symlink for bundled APEX
For bundled APEXes, files that are included in the APEX due to
transitive dependencies are replaced with symlinks to the same files in
the system partition. For example, /apex/com.android.art/lib/libbase.so
is a symlink to /system/lib/libbase.so. This is to reduce the
storage/ram overhead of APEXes.

Note that this optimization is done only for bundled APEXes (i.e. APEXes
that are built "WITH" the platform and pre-installed). Unbundled APEXes
(i.e., APEXes for Mainline modules built/signed by Google) are not
affected.

Bug: 144533348
Test: m (test amended)
Change-Id: I1ed21b1d47d93d2b8c296f8f227105031b29afed