Extract the osTypeSpecificInfo code from module creation loop
Extract the functionality to create an osTypeSpecificInfo struct,
to optimize the properties, and add its properties to a property set
into methods of the *osTypeSpecificInfo struct.

This change is in preparation for adding support for link type which
is another dimension within arch type which itself sits within os type.

Bug: 153306490
Test: m nothing
Bug: 142918168
Merged-In: I025ee90e1461f7389bf4a9d056b281453068cf87
Change-Id: I025ee90e1461f7389bf4a9d056b281453068cf87