Update patch set 1
Patch Set 1:

Don't know if anyone else is interested in topic:ten-build-nice but it bugs the crap out of me that build defaults to nice -19 (I already have things running at 19) and I couldn't find any way to change it so I did this.

If build is started with nice 0, it drops to 19 preserving the existing behaviour.
If build is started with nice != 0, it will continue to use the existing level and no longer drop to 19.

Could just keep these around for picks.

Patch-set: 1
Reviewer: Gerrit User 3075 <3075@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>