Retry: "java_sdk_library: Do not expose stubs implementation jar"
The stubs header jar is optimized for use as a dependency for others
to use. It only changes if there is a significant difference in the
externals of the classes, i.e. anything that a library being compiled
against depends upon. So changes to implementations of method or the
addition/removal of private methods, fields will have no impact.

As there is no benefit in returning the implementation of the stubs
jar this change removes it. The implementation is still used when
taking a snapshot as the header jar is an internal build artefact
that is not suitable for long term snapshot.

Bug: 155164730
Test: m droid
(cherry picked from commit 23970f4285f2ba70b5a35a56a6dffa9afc1f37d9)
Change-Id: I8277ec643837514d74cb57ad4f236ceb1f5c6c5b